refactor(canvas): extract skia_canvas_mask.cpp (R2-3 first cut)
skia_canvas.cpp was 3,586 lines with 40 touches in the last 60 days
— the hottest core/canvas file by far. Pulling the CSS mask-image /
mask-size paint slice off the bottom of the file into its own TU
isolates mask-specific edits.

The mask slice (pulp #1737 / #1515) handles linear-gradient and
radial-gradient mask shapes; url() and other shapes fall through to
the no-mask path. CSS Masking Module Level 1 §5's kDstIn composite
is documented at the SkiaCanvas::restore() call site (which stays
in skia_canvas.cpp).

Changes:
* New `core/canvas/src/skia_canvas_mask.cpp` (~367 lines incl.
  the Skia include preamble + skia_canvas.hpp). Preamble mirrors
  skia_canvas.cpp's ordering (Skia headers BEFORE skia_canvas.hpp
  per the C++ name-lookup rule documented at the top of the
  original file).
* `skia_canvas.cpp` drops from 3,586 → 3,282 lines (-8%). A
  marker comment at the original section header points readers at
  the new TU.
* `core/canvas/CMakeLists.txt` adds the new source under the
  PULP_HAS_SKIA guard so non-Skia builds stay unaffected.

Validation per Codex R2-3 risk callout: the visual harness has Yoga
semantic goldens, not Skia raster goldens, so "renders identical"
isn't testable as written. Substitute byte-stable verification:
* pulp-canvas builds clean (Release with Skia linked).
* The mask code is structurally identical — every line moved
  verbatim into the new TU.
* No public API surface change in pulp/canvas/skia_canvas.hpp.

This is item 1 of Phase 4. Future cuts on skia_canvas.cpp (path /
text / image / transform / gradients / box-shadow / opacity-layers
/ GPU SDF / waveform / SkSL shaders) follow the same pattern. The
4,670-line `core/view/src/widget_bridge.cpp` (the eventual A1 full
split) is the other major core/* monolith awaiting splits.
