+++ core/canvas/CMakeLists.txt
@@ -57,6 +57,7 @@ target_link_libraries(pulp-canvas PUBLIC pulp::runtime)
 if(PULP_HAS_SKIA)
     target_sources(pulp-canvas PRIVATE
         src/skia_canvas.cpp
+        src/skia_canvas_mask.cpp
         src/bundled_fonts.cpp
     )
